requirements.txt
altair==5.5.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
dotenv==0.9.9
executing==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.72.1
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
narwhals==1.41.1
nest-asyncio==1.6.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
pywin32==309
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
rsa==4.9.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.38
stack-data==0.6.3
streamlit==1.45.1
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.2.0
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
